+++ src/shared/types/margin-profit.types.ts
@@ -1,7 +1,12 @@
-import { number, object, ObjectSchema } from "yup";
+import { QuoteCourier, QuoteSource } from "./quotes.types";
 
 export type ProfitMarginType = 'percentage' | 'absolute'
 
+export type MarginProfitSubscreens = 'view' | 'edit'
+
+/**
+ * Type used for the profit margin form selection input
+ */
 export type ProfitMarginTypeOption = {
   label: string;
   value: ProfitMarginType;
@@ -12,10 +17,20 @@ export interface ProfitMargin {
   type: ProfitMarginType
 }
 
+export interface CourierGlobalConfig {
+  name: QuoteCourier;
+  profitMargin: ProfitMargin;
+}
+
+export interface ProviderGlobalConfig {
+  name: QuoteSource;
+  couriers: CourierGlobalConfig[];
+}
+
 
 export interface MarginProfitResponse {
   data: {
-    profitMargin: ProfitMargin
+    providers: ProviderGlobalConfig[]
   }
   error: null;
   message: null;
@@ -25,16 +40,12 @@ export interface MarginProfitResponse {
 }
 
 export type UpdateMarginProfitPayload = {
-  profitMargin: {
-    value: number
-    type: ProfitMarginType
-  }
+  providers: ProviderGlobalConfig[]
 }
 
-export type MarginProfitForm = {
-  value: number
+export type CourierForm = {
+  id: string
+  value: number | null;
+  courier: QuoteCourier;
+  profitMarginType: ProfitMarginTypeOption;
 }
-
-export const MarginProfitSchema: ObjectSchema<MarginProfitForm> = object().shape({
-  value: number().required('Por favor, ingrese el valor del margen de ganancia').min(1, 'El valor debe ser mayor que 0'),
-})