+++ src/shared/ui/atoms/ErrorMessage.tsx
@@ -2,10 +2,11 @@ import { ReactNode } from "react";
 
 interface ErrorMessageProps {
   children: ReactNode;
+  className?: string;
 }
 
-export const ErrorMessage = ({ children }: ErrorMessageProps) => {
+export const ErrorMessage = ({ children, className }: ErrorMessageProps) => {
   return (
-    <p className="text-red-500 text-sm mt-1">{children}</p>
+    <p className={`text-red-500 text-sm mt-1 ${className}`}>{children}</p>
   )
 }